TIP-1287: move listeners
ConfigureCategoryTreeForExportJobsAfterChangingTheChannelCategoryTree
InitCompletenessDbSchemaSubscriber
ComputeEntityRawValuesSubscriber
LoadEntityWithValuesSubscriber
