+++ src/Akeneo/Pim/Enrichment/Bundle/EventSubscriber/EntityWithValues/ComputeEntityRawValuesSubscriber.php
@@ -23,7 +23,7 @@ use Symfony\Component\Serializer\Normalizer\NormalizerInterface;
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ComputeEntityRawValuesSubscriber implements EventSubscriberInterface
+final class ComputeEntityRawValuesSubscriber implements EventSubscriberInterface
 {
     /** @var NormalizerInterface */
     protected $normalizer;
