+++ src/Akeneo/Pim/Enrichment/Bundle/EventSubscriber/Family/ComputeCompletenessOnFamilyUpdateSubscriber.php
@@ -22,7 +22,7 @@ use Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInt
  * @copyright 2017 Akeneo SAS (http://www.akeneo.com)
  * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
  */
-class ComputeCompletenessOnFamilyUpdateSubscriber implements EventSubscriberInterface
+final class ComputeCompletenessOnFamilyUpdateSubscriber implements EventSubscriberInterface
 {
     /** @var TokenStorageInterface */
     private $tokenStorage;
